+++ StateObject/BalanceState/balancestate.cc
@@ -15,14 +15,14 @@ void BalanceState::writeJson(Writer<StringBuffer>& writer) const
     writer.String("offsets");
     writer.StartObject();
     {
-      writer.String("hipRollR").Double(d_offsets.hipRollR);
-      writer.String("hipRollL").Double(d_offsets.hipRollL);
-      writer.String("kneeR").Double(d_offsets.kneeR);
-      writer.String("kneeL").Double(d_offsets.kneeL);
-      writer.String("anklePitchR").Double(d_offsets.anklePitchR);
-      writer.String("anklePitchL").Double(d_offsets.anklePitchL);
-      writer.String("ankleRollR").Double(d_offsets.ankleRollR);
-      writer.String("ankleRollL").Double(d_offsets.ankleRollL);
+      writer.String("hipRollR").Int(d_offsets.hipRollR);
+      writer.String("hipRollL").Int(d_offsets.hipRollL);
+      writer.String("kneeR").Int(d_offsets.kneeR);
+      writer.String("kneeL").Int(d_offsets.kneeL);
+      writer.String("anklePitchR").Int(d_offsets.anklePitchR);
+      writer.String("anklePitchL").Int(d_offsets.anklePitchL);
+      writer.String("ankleRollR").Int(d_offsets.ankleRollR);
+      writer.String("ankleRollL").Int(d_offsets.ankleRollL);
     }
     writer.EndObject();
   }
